Add a basic text interface to GUI program
The GUI program (which does NOT incorporate a graphical interface, but
is the placeholder for the graphical interface to be developed), now
includes a text interface.

On each loop, the program will prompt the user to enter a command.
Correct commands have the form "command [value]", but this is not yet
enforced. In this version, the value updates the "desired current", but
this will shortly be altered to reflect some actual slow control uses.

Important note: this code does not incorporate multithreading. This
means that while the program waits for user input, it blocks and cannot
receive updated values from the network. It can only update one packet
at a time each time the enter key is pressed. This is a serious flaw
which, as far as I know, cannot be avoided by a single-thread program.