Add GUI command to report FEE voltages
The 'v' command now obtains the FEE housekeeping voltages from the Pi and prints them to the "GUI" interface.

In this commit, the Pi program does not yet implement the low-level commands to read these data from the backplane, but that (untestable) functionality will be added next.

To allow this and similar functionality going forward, the backplane
class and protocol buffers have finally been rewritten to incorporate
their actual variables. The data buffer and backplane class now include
arrays for storing the voltages for each FEE, and the settings buffer
and backplane class have a variable encoding what data (if any) is
requested. As of this commit the only possibility is the FEE voltages,
but the intention is that future updates will more types of data
depending on what's needed.

This commit also fixes a significant bug. Previously, the
poll_connections() function did not enforce a default value when
determining whether sockets were ready for receiving and sending. As a
result, once the GUI sent a single message to the Pi, the Pi would keep
treating the connection as having sent a message on every loop and
updating its data accordingly. Now, socket status is reset to MSG_STANDBY
whenever poll_connections() is called, preventing this issue.