Include low level SPI communication
This commit introduces low level SPI communication from the Pi to the
(physical) backplane. The code for doing this is copied and pasted from
the existing BP SPI interface code, which minor modifications for
integration with the rest of the code and for style.

This code does not compile because the necessary header file bcm2835.h
is missing. Irrespective of that, the code cannot run or be tested until
the backplane is built and operational.