Create simulation mode for testing
When the pi program cannot initiate SPI communication (e.g. because the
code is running on a computer not connected to a backplane), the code
will now run in simulation mode instead of exiting with an error.
Instead of sending SPI commands, the pi program will generate simple fake
data to send to the GUI. This is useful for testing on a laptop without
having to upload the code to the actual functioning raspberry pi.