Display voltage data in gui program
The full functionality of the 'v' command in the bp_test_pi program is
now implemented. A bug in which the voltages stored in the backplane
object on the pi were not copied into the data buffer has been fixed.
Previously, the data were obtained correctly, but only an empty data
buffer was sent over the network for display. Also, a time delay has
been introduced into the gui program to give time for the pi to read and
send values before displaying updates to the user.

Finally, the text output of the voltages has been restructured to match
that provided by the bp_test_pi program.