Add d command to set trigger at time
The GUI can now send up to four SPI words to the Pi and receive up to 11
SPI words of data back for displaying to the user. This replicates the
behavior of the bp_test_pi program.

A strange bug results when printing the SPI data, in which cin receives
an extra newline character. This is solved by simply clearing cin after
printing SPI data, but I don't know what the cause of the problem is.