Log messages sent through the server
The skeleton of a basic logging functionality is implemented. Whenever a
message is sent through the server, a function is called which can
access the content of the message (data or settings) and use that
information to do something. Currently a short message is printed to the
screen.

A better system would log whenever a message is received by the server,
not just when it's transmitted to its destination, since right now no
data would be logged if eg the user interface is closed. This will
require modifying the server code in sc_network.