Hide error messages in network code
An error message will no longer be printed when a connection cannot be
established on a socket; instead an invalid socket will be returned as
normal, without any printout to the screen. This is relevant when the pi
tries to connect to the server, which is not running. It will try in a
loop, resulting in a continual stream of useless error messages which
will overwhelm a log file. This change allows for saving the output of
the pi program to a text file.