Add target module controller program
A new program is added to handle commands for the target modules in a
way analoguous to the pi's control of the backplane. The current
slow_control_tm.cc program is a copy of the pi program with no
functionality.

The networking functions have been modified to allow the server to
communicate with both the tm controller and the pi. The modifications
are relatively minor, with the only exception being the functionality
for the server in the update_network() function. Variables from the pi
and tm controller are both sent to the gui as before, but gui commands
are not passed forward. This will allow the server to break down the
high level commands from the gui into low level commands for the
controller programs.