Add new backend library for target module control
This commit replaces the old sc_backplane and sc_logistics libraries
with a set of control libraries for each of the four programs in the
slow control setup: the interface, server, backplane controller (i.e.
pi) and the target module controller. The target module controller is a
separate program that uses TargetDriver to talk to the modules without
necessarily including high level logic, in a way analagous to how the
current pi program operates.

Each of the four programs contains a class for performing its specific
functions. These classes all inherit from a single RunControl class
that contains fundamental high level information and functionality.

It's intended that these classes will be wrapped as a Python library
using SWIG. Then python scripts will be able to call these functions to
set up the network and to interpret, perform, and log commands. This
will allow us to make use of both this C++ codebase and the existing
python target module control and data taking scripts without having to
rewrite either in a new language.