Refactor server and pi code to be wrapped library
The server program has been changed from a hub that passively sends
messages back and forth from the client programs to an active run
control program that parses high level commands and transmits the
corresponding low level commands.

On startup, the server reads a configuration file containing the
structure of all high level and low level commands and stores this
information in memory. By storing this information in a configuration
file, new commands can be easily added in testing by users. At a later
date the commands can be "frozen in" into a header file.

Each program (server, pi, tm, interface) has its corresponding control
class. Previously, each inherited from a parent class; this proved
unecessary and bad style, and so has been removed. The control classes
present an API that is meant to be used either natively in C++ or
wrapped in Python within the main loop of the corresponding programs.